+++ CHANGELOG.md
@@ -3,11 +3,13 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+* `flexberry-groupedit` component:
+    * Add check all at page button.
+
 ### Changed
 * `flexberry-lookup` component:
     * Dropdown mode uses projection from `projection` property.
-* `flexberry-groupedit` component:
-    * Replaced check all at all pages button on check all at page.
 
 ## [2.1.0-beta.3] - 2019-03-15
 ### Added